pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "abias-sgv"
version = "0.1.0"
description = "Let's Think in Two Steps: Mitigating Agreement Bias in MLLMs with Self-Grounded Verification"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Dependencies for core_utils
    "numpy",
    "pandas",
    "pyyaml",
    "pillow",
    "pathvalidate",
    "pydantic",
    "requests",
    "filelock",
    
    # Dependencies for llms
    "aiolimiter",
    "anthropic",
    "google-genai",
    "openai",
    "torch",
    "torchvision",
    "torchaudio",
    "transformers",
    "qwen_vl_utils",
    "tiktoken",
    "accelerate",
    "fastapi",
    "langdetect",
    "ftfy"
    # bitsandbytes (optional)
    # flash_attn (optional)
    # vllm (optional)
]

[project.optional-dependencies]
vwa = [
    # VisualWebArena specific dependencies
    "gymnasium",
    "playwright",
    "matplotlib",
    "evaluate",
    "beartype",
    "nltk",
    "scikit-image",
    "fastapi",
    "python-multipart",
    "uvicorn",
    "docker",
    "flask",
    "bs4"
]
osw = [
    # OSWorld specific dependencies
    "fabric",
    "opencv-python",
    "matplotlib",
    "pynput",
    "pyautogui",
    "psutil",
    "requests-toolbelt",
    "lxml",
    "cssselect",
    "xmltodict",
    "openpyxl",
    "python-docx",
    "python-pptx",
    "pypdf",
    "PyGetWindow",
    "rapidfuzz",
    "pyacoustid",
    "pygame",
    "opencv-python",
    "ImageHash",
    "scikit-image",
    "librosa",
    "pymupdf",
    "chardet",
    "playwright",
    "backoff",
    "formulas",
    "pydrive",
    "fastdtw",
    "odfpy",
    "func-timeout",
    "beautifulsoup4",
    "dashscope",
    "PyYaml",
    "mutagen",
    "easyocr",
    "borb",
    "pypdf2",
    "pdfplumber",
    "wandb",
    "wrapt_timeout_decorator",
    "gdown",
    "tiktoken",
    "groq",
    "boto3",
    "azure-identity",
    "azure-mgmt-compute",
    "azure-mgmt-network",
    "docker",
    "loguru",
    "bs4"
]
offline=[
    "scikit-learn",
    "scipy"
]
robomimic = [
# Under construction
]

dev = [
    "pytest",
    "isort",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "core_utils*",
    "llms*",
]
exclude = [
    "agrb*",
    "vwa*",
    "osw*",
    "offline_experiments*",
    "robomimic*",
    "data*",
    "*.tests*",
    "*.__pycache__*",
]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.txt", "*.md"]
